feat: resume page enhancements
